Turned anime card into a component
